modified:   .next/build-manifest.json 	modified:   .next/server/middleware-build-manifest.js 	modified:   .next/static/development/_buildManifest.js 	modified:   pages/api/hotels/[id].js 	modified:   pages/api/hotels/availability.js 	modified:   pages/api/hotels/index.js 	modified:   pages/api/hotels/search.js

+++ pages/api/hotels/[id].js
@@ -1,10 +1,12 @@
 // pages/api/hotels/[id].js
 import { ObjectId } from 'mongodb';
 import clientPromise from '../../../lib/mongodb';
+import hotelsData from '../../../data/hotels.json';
+import { DEMO_MODE } from '../../../lib/constants';
 
 export default async function handler(req, res) {
-  const client = await clientPromise;
-  const db = client.db();
+  const client = DEMO_MODE ? null : await clientPromise;
+  const db = DEMO_MODE ? null : client.db();
   const { id } = req.query;
 
   if (!ObjectId.isValid(id)) {
@@ -16,6 +18,11 @@ export default async function handler(req, res) {
   switch (req.method) {
     case 'GET':
       try {
+        if (DEMO_MODE) {
+          const hotel = hotelsData.find(h => (h._id || h.id) == id);
+          if (!hotel) return res.status(404).json({ error: 'Hotel not found' });
+          return res.status(200).json(hotel);
+        }
         const hotel = await hotels.findOne({ _id: new ObjectId(id) });
         if (!hotel) return res.status(404).json({ error: 'Hotel not found' });
         return res.status(200).json(hotel);
@@ -26,6 +33,9 @@ export default async function handler(req, res) {
 
     case 'PUT':
       try {
+        if (DEMO_MODE) {
+          return res.status(200).json({ message: 'Demo mode: Not persisted' });
+        }
         const updateData = req.body;
         await hotels.updateOne(
           { _id: new ObjectId(id) },
@@ -40,6 +50,9 @@ export default async function handler(req, res) {
 
     case 'DELETE':
       try {
+        if (DEMO_MODE) {
+          return res.status(200).json({ message: 'Demo mode: Not persisted' });
+        }
         await hotels.deleteOne({ _id: new ObjectId(id) });
         return res.status(204).end();
       } catch (err) {
